+++ internal/graph/lock_unix.go
@@ -1,0 +1,23 @@
+//go:build !windows
+
+package graph
+
+import (
+	"fmt"
+	"os"
+	"syscall"
+)
+
+func lockFile(f *os.File) error {
+	if err := syscall.Flock(int(f.Fd()), syscall.LOCK_EX|syscall.LOCK_NB); err != nil {
+		if err == syscall.EWOULDBLOCK {
+			return ErrGraphLocked
+		}
+		return fmt.Errorf("acquiring graph lock: %w", err)
+	}
+	return nil
+}
+
+func unlockFile(f *os.File) error {
+	return syscall.Flock(int(f.Fd()), syscall.LOCK_UN)
+}